# Module import optimizations to git-meta