--- a/node/lib/git-meta.js
+++ b/node/lib/git-meta.js
@@ -69,11 +69,9 @@ const version      = require("./cmd/version");
  * @param {ArgumentParser} parser
  * @param {String}         commandName
  * @param {Object}         module
- * @param {Function}       module.configureParser
- * @param {Function}       module.executeableSubcommand
- * @param {String}         module.helpText
+ * @param {Boolean}        skipNodeGit
  */
-function configureSubcommand(parser, commandName, module) {
+function configureSubcommand(parser, commandName, module, skipNodeGit) {
     const subParser = parser.addParser(commandName, {
         help: module.helpText,
         description: module.description,
@@ -81,11 +79,17 @@ function configureSubcommand(parser, commandName, module) {
     module.configureParser(subParser);
     subParser.setDefaults({
         func: function (args) {
-            const NodeGit = require("nodegit");
-            // see https://github.com/nodegit/nodegit/issues/827 -- this is required
-            // to prevent random hard crashes with e.g. parallelism in index operations.
-            // Eventually, this will be nodegit's default.
-            NodeGit.setThreadSafetyStatus(NodeGit.THREAD_SAFETY.ENABLED_FOR_ASYNC_ONLY);
+            // This optimization allows to skip importing NodeGit
+            // if not required (eg. version).
+            if (!skipNodeGit) {
+                const NodeGit = require("nodegit");
+                // see https://github.com/nodegit/nodegit/issues/827 -- this is
+                // required to prevent random hard crashes with e.g.
+                // parallelism  in index operations. Eventually, this will be
+                // nodegit's default.
+                NodeGit.setThreadSafetyStatus(
+                    NodeGit.THREAD_SAFETY.ENABLED_FOR_ASYNC_ONLY);
+            }
 
             module.executeableSubcommand(args)
             .catch(function (error) {
@@ -147,11 +151,16 @@ const commands = {
     "version": version,
 };
 
+// These optimized commands do not require NodeGit, and can skip importing it.
+const optimized = {
+    "version": true,
+};
+
 // Configure the parser with commands in alphabetical order.
 
 Object.keys(commands).sort().forEach(name => {
     const cmd = commands[name];
-    configureSubcommand(subParser, name, cmd);
+    configureSubcommand(subParser, name, cmd, (null !== optimized[name]));
 });
 
 const blacklist = new Set([