**NodeGit is slow.** 
These module import optimizations result in some improved performance in specific targeted areas where interactive performance is useful. 

Performance changes (mid-2012 MBA w/ SSD)
* **git meta root, 700ms -> 260ms (-63%)**
* --help (and other strictly argparse-only operations), 650ms -> 230ms (-65%)
* forwarded commands to pure-git all improved. eg.
* * log 650ms -> 300ms (-54%)
* * diff 700ms -> 270ms (-62%)

_Highly recommend reviewing this commit-by-commit._